# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RS Mas Amsyar WEB API",
        "description": "\r\n * > Footprint 1.0 (2024 June 25) `Developed along with the web frontend to be consumed by and also enabled integration with 3rd party API in the future`\r\n * > Footprint x (2024 xx xx) `this-is-footnote-format-please-continue-on-next-dev`",
        "contact": {
            "name": "Evelline Krist.",
            "email": "ev.attoff@gmail.com"
        },
        "version": "VER 1.0",
        "x-logo": {
            "url": "https://upload.wikimedia.org/wikipedia/commons/thumb/9/9a/Laravel.svg/985px-Laravel.svg.png"
        }
    },
    "paths": {
...
        "/api/page": {
            "get": {
                "tags": [
                    "Page"
                ],
                "summary": "Display a listing of items",
                "operationId": "pageIndex",
                "parameters": [
                    {
                        "name": "_page",
                        "in": "query",
                        "description": "current page",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    },
                    {
                        "name": "_limit",
                        "in": "query",
                        "description": "max item in a page",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 10
                        }
                    },
                    {
                        "name": "_search",
                        "in": "query",
                        "description": "word to search",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "_dir",
                        "in": "query",
                        "description": "order by direction",
                        "required": false,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Page"
                ],
                "summary": "Store a newly created item",
                "operationId": "pageStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "slug",
                                    "title",
                                    "layout"
                                ],
                                "properties": {
                                    "slug": {
                                        "type": "string",
                                        "example": "clean-eating"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Clean Eating"
                                    },
                                    "layout": {
                                        "type": "string",
                                        "example": "img_body"
                                    },
                                    "img_main": {
                                        "description": "Image to upload",
                                        "type": "file"
                                    },
                                    "file_main": {
                                        "description": "File to upload",
                                        "type": "file"
                                    },
                                    "file_link": {
                                        "type": "string"
                                    },
                                    "body": {
                                        "type": "string",
                                        "example": "<p>Menjadi sehat adalah impian semua orang. Makanan yang selama ini kita pikir sehat ternyata belum tentu <i>`sehat`</i> bagi tubuh kita.</p>"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "201": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport_token_ready": [],
                        "passport": []
                    }
                ]
            }
        },
        "/api/page/{id}": {
            "get": {
                "tags": [
                    "Page"
                ],
                "summary": "Display the specified item",
                "operationId": "pageShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifier of item that needs to be displayed",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Item not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Page"
                ],
                "summary": "Update the specified item",
                "operationId": "pageUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifier of item that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "clean-eating"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "layout"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "[EDITED] Clean Eating"
                                    },
                                    "layout": {
                                        "type": "string",
                                        "example": "img_body"
                                    },
                                    "img_main": {
                                        "description": "Image to upload",
                                        "type": "file"
                                    },
                                    "file_main": {
                                        "description": "File to upload",
                                        "type": "file"
                                    },
                                    "file_link": {
                                        "type": "string"
                                    },
                                    "body": {
                                        "type": "string",
                                        "example": "<p>[EDITED]</p><p>Menjadi sehat adalah impian semua orang. Makanan yang selama ini kita pikir sehat ternyata belum tentu <i>`sehat`</i> bagi tubuh kita.</p>"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Item not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport_token_ready": [],
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Page"
                ],
                "summary": "Remove the specified item",
                "operationId": "pageDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifier of item that needs to be removed",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Item not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport_token_ready": [],
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "passport_token_ready": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format without \"Bearer \"",
                "name": "Authorization",
                "in": "header"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}